[Started] Moved exception.dart to utils folder
