+++ lib/src/jiffy.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:intl/intl.dart';
-import 'package:jiffy/src/exception/exception.dart';
+import 'package:jiffy/src/utils/exception.dart';
 import 'package:jiffy/src/relative_time/relative_time.dart' as relative;
 import 'package:jiffy/src/utils/normalize_units.dart';
 import 'package:jiffy/src/utils/ordinalLocale.dart';
