+++ lib/src/utils/exception.dart
